Begin experiments
