+++ tests_unit/test_cli.py
@@ -1,0 +1,5 @@
+import unittest
+
+
+class CLITests(unittest.TestCase):
+    pass
